requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
cloudinary
python-dotenv
psycopg2-binary
python-slugify
groq
